Split namespace component in 3 files
The component, the model and the view.
