+++ gaphor/ui/namespace.py
@@ -10,24 +10,21 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING, Optional
 
-from gi.repository import Gdk, Gio, GLib, Gtk, Pango
+from gi.repository import Gdk, Gio, GLib, Gtk
 
 from gaphor import UML
 from gaphor.core import action, event_handler, gettext, transactional
-from gaphor.core.format import format, parse
-from gaphor.core.modeling import (
-    AttributeUpdated,
-    DerivedSet,
-    Diagram,
-    ElementCreated,
-    ElementDeleted,
-    ModelFlushed,
-    ModelReady,
-)
+from gaphor.core.format import format
+from gaphor.core.modeling import Diagram
 from gaphor.ui.abc import UIComponent
 from gaphor.ui.actiongroup import create_action_group
 from gaphor.ui.event import DiagramOpened
-from gaphor.ui.iconname import get_icon_name
+from gaphor.ui.namespacemodel import (
+    RELATIONSHIPS,
+    NamespaceModel,
+    NamespaceModelRefreshed,
+)
+from gaphor.ui.namespaceview import NamespaceView
 
 if TYPE_CHECKING:
     from gaphor.core.eventmanager import EventManager
@@ -35,416 +32,6 @@ if TYPE_CHECKING:
 
 
 log = logging.getLogger(__name__)
-
-
-class RELATIONSHIPS:
-    name = "0"
-    owner = None
-
-
-def relationship_iter(model, iter):
-    if iter is None or isinstance(model.get_value(iter, 0), UML.Package):
-        child_iter = model.iter_children(iter)
-        while child_iter:
-            maybe_relationships = model.get_value(child_iter, 0)
-            if maybe_relationships is RELATIONSHIPS:
-                return child_iter
-            child_iter = model.iter_next(child_iter)
-        return model.append(iter, [RELATIONSHIPS])
-    else:
-        return iter
-
-
-def relationship_iter_parent(model, iter):
-    while model.get_value(iter, 0) is RELATIONSHIPS:
-        iter = model.iter_parent(iter)
-    return iter
-
-
-class NamespaceView(Gtk.TreeView):
-
-    TARGET_STRING = 0
-    TARGET_ELEMENT_ID = 1
-    DND_TARGETS = [
-        Gtk.TargetEntry.new("STRING", 0, TARGET_STRING),
-        Gtk.TargetEntry.new("text/plain", 0, TARGET_STRING),
-        Gtk.TargetEntry.new("gaphor/element-id", 0, TARGET_ELEMENT_ID),
-    ]
-
-    def __init__(self, model: Gtk.TreeModel, element_factory: ElementFactory):
-        Gtk.TreeView.__init__(self)
-        self.set_model(model)
-        self.element_factory = element_factory
-
-        self.set_property("headers-visible", False)
-        self.set_property("search-column", 0)
-
-        selection = self.get_selection()
-        selection.set_mode(Gtk.SelectionMode.BROWSE)
-        column = Gtk.TreeViewColumn.new()
-        # First cell in the column is for an image...
-        cell = Gtk.CellRendererPixbuf()
-        column.pack_start(cell, 0)
-        column.set_cell_data_func(cell, self._set_pixbuf, None)
-
-        # Second cell if for the name of the object...
-        cell = Gtk.CellRendererText()
-        cell.connect("edited", self._text_edited)
-        column.pack_start(cell, 0)
-        column.set_cell_data_func(cell, self._set_text, None)
-
-        self.append_column(column)
-
-        # drag
-        self.drag_source_set(
-            Gdk.ModifierType.BUTTON1_MASK | Gdk.ModifierType.BUTTON3_MASK,
-            NamespaceView.DND_TARGETS,
-            Gdk.DragAction.COPY | Gdk.DragAction.MOVE,
-        )
-        self.connect("drag-data-get", NamespaceView.on_drag_data_get)
-        self.connect("drag-data-delete", NamespaceView.on_drag_data_delete)
-
-        # drop
-        self.drag_dest_set(
-            Gtk.DestDefaults.ALL, [NamespaceView.DND_TARGETS[-1]], Gdk.DragAction.MOVE
-        )
-        self.connect("drag-motion", NamespaceView.on_drag_motion)
-        self.connect("drag-data-received", NamespaceView.on_drag_data_received)
-
-    def get_selected_element(self):
-        selection = self.get_selection()
-        model, iter = selection.get_selected()
-        if not iter:
-            return
-        return model.get_value(iter, 0)
-
-    def expand_root_nodes(self):
-        self.expand_row(path=Gtk.TreePath.new_first(), open_all=False)
-
-    def _set_pixbuf(self, column, cell, model, iter, data):
-        element = model.get_value(iter, 0)
-
-        if element is RELATIONSHIPS:
-            cell.set_property("icon-name", None)
-            cell.set_property("visible", False)
-        elif isinstance(element, (UML.Property, UML.Operation)):
-            cell.set_property("icon-name", None)
-            cell.set_property("visible", False)
-        else:
-            icon_name = get_icon_name(element)
-            cell.set_property("visible", True)
-
-            if icon_name:
-                cell.set_property("icon-name", icon_name)
-
-    def _set_text(self, column, cell, model, iter, data):
-        """Set font and of model elements in tree view."""
-        element = model.get_value(iter, 0)
-
-        cell.set_property(
-            "weight",
-            Pango.Weight.BOLD if isinstance(element, Diagram) else Pango.Weight.NORMAL,
-        )
-
-        cell.set_property(
-            "style",
-            Pango.Style.ITALIC
-            if isinstance(element, (UML.Classifier, UML.BehavioralFeature))
-            and element.isAbstract
-            else Pango.Style.NORMAL,
-        )
-
-        if element is RELATIONSHIPS:
-            text = gettext("<Relationships>")
-        else:
-            text = format(element) or "<None>"
-        cell.set_property("text", text)
-
-    @transactional
-    def _text_edited(self, cell, path_str, new_text):
-        """The text has been edited.
-
-        This method updates the data object. Note that 'path_str' is a
-        string where the fields are separated by colons ':', like this:
-        '0:1:1'. We first turn them into a tuple.
-        """
-        model = self.get_property("model")
-        iter = model.get_iter_from_string(path_str)
-        element = model.get_value(iter, 0)
-        try:
-            parse(element, new_text)
-        except TypeError:
-            log.debug(f"No parser for {element}")
-
-    def on_drag_data_get(self, context, selection_data, info, time):
-        """Get the data to be dropped by on_drag_data_received().
-
-        We send the id of the dragged element.
-        """
-        selection = self.get_selection()
-        model, iter = selection.get_selected()
-        if iter:
-            element = model.get_value(iter, 0)
-            if info == NamespaceView.TARGET_ELEMENT_ID:
-                selection_data.set(
-                    selection_data.get_target(), 8, str(element.id).encode()
-                )
-            else:
-                selection_data.set(
-                    selection_data.get_target(), 8, str(element.name).encode()
-                )
-        return True
-
-    def on_drag_data_delete(self, context):
-        """Delete data from original site, when `ACTION_MOVE` is used."""
-        self.emit_stop_by_name("drag-data-delete")
-
-    # Drop
-    def on_drag_motion(self, context, x, y, time):
-        path_pos_or_none = self.get_dest_row_at_pos(x, y)
-        if path_pos_or_none:
-            self.set_drag_dest_row(*path_pos_or_none)
-        else:
-            self.set_drag_dest_row(
-                Gtk.TreePath.new_from_indices([len(self.get_model()) - 1]),
-                Gtk.TreeViewDropPosition.AFTER,
-            )
-        return True
-
-    @transactional
-    def on_drag_data_received(self, context, x, y, selection, info, time):
-        """Drop the data send by on_drag_data_get()."""
-        self.stop_emission_by_name("drag-data-received")
-        if info == NamespaceView.TARGET_ELEMENT_ID:
-            element_id = selection.get_data().decode()
-            drop_info = self.get_dest_row_at_pos(x, y)
-        else:
-            drop_info = None
-
-        if drop_info:
-            model = self.get_model()
-            element = self.element_factory.lookup(element_id)
-            if not isinstance(element, (Diagram, UML.Package, UML.Type)):
-                log.debug("Element {element} can not be dropped")
-                return context.finish(False, False, time)
-            path, position = drop_info
-            iter = model.get_iter(path)
-            dest_element = model.get_value(iter, 0)
-            assert dest_element
-            if dest_element is RELATIONSHIPS:
-                iter = relationship_iter_parent(model, iter)
-                dest_element = model.get_value(iter, 0)
-
-            # Add the item to the parent if it is dropped on the same level,
-            # else add it to the item.
-            if position in (
-                Gtk.TreeViewDropPosition.BEFORE,
-                Gtk.TreeViewDropPosition.AFTER,
-            ):
-                parent_iter = model.iter_parent(iter)
-                dest_element = (
-                    None if parent_iter is None else model.get_value(parent_iter, 0)
-                )
-            try:
-                # Check if element is part of the namespace of dest_element:
-                ns = dest_element
-                while ns:
-                    if ns is element:
-                        raise AttributeError
-                    ns = ns.namespace
-
-                # Set package. This only works for classifiers, packages and
-                # diagrams. Properties and operations should not be moved.
-                if dest_element is None:
-                    del element.package
-                else:
-                    element.package = dest_element
-
-            except AttributeError as e:
-                log.info(f"Unable to drop data {e}")
-                context.finish(False, False, time)
-            else:
-                context.finish(True, True, time)
-
-
-class NamespaceModelRefreshed:
-    def __init__(self, model):
-        self.model = model
-
-
-class NamespaceModel:
-    def __init__(self, event_manager: EventManager, element_factory: ElementFactory):
-        self.event_manager = event_manager
-        self.element_factory = element_factory
-        self.model = Gtk.TreeStore.new([object])
-
-        event_manager.subscribe(self.refresh)
-        event_manager.subscribe(self._on_element_create)
-        event_manager.subscribe(self._on_element_delete)
-        event_manager.subscribe(self._on_association_set)
-        event_manager.subscribe(self._on_attribute_change)
-
-    def shutdown(self):
-        em = self.event_manager
-        em.unsubscribe(self.refresh)
-        em.unsubscribe(self._on_element_create)
-        em.unsubscribe(self._on_element_delete)
-        em.unsubscribe(self._on_association_set)
-        em.unsubscribe(self._on_attribute_change)
-
-    def sorted(self):
-        """Get a sorted version of this model."""
-        sorted_model = Gtk.TreeModelSort(model=self.model)
-
-        def sort_func(model, iter_a, iter_b, userdata):
-            va = model.get_value(iter_a, 0)
-            vb = model.get_value(iter_b, 0)
-
-            # Put Relationships pseudo-node at top
-            if va is RELATIONSHIPS:
-                return -1
-            if vb is RELATIONSHIPS:
-                return 1
-
-            a = (format(va) or "").lower()
-            b = (format(vb) or "").lower()
-            if a == b:
-                return 0
-            if a > b:
-                return 1
-            return -1
-
-        sorted_model.set_sort_func(0, sort_func, None)
-        sorted_model.set_sort_column_id(0, Gtk.SortType.ASCENDING)
-
-        return sorted_model
-
-    def iter_children(self, iter):
-        return self.model.iter_children(iter)
-
-    def iter_n_children(self, iter):
-        return self.model.iter_n_children(iter)
-
-    def get_element(self, iter):
-        return self.model.get_value(iter, 0)
-
-    def iter_for_element(self, element, old_owner=0):
-        """Get the Gtk.TreeIter for an element in the Namespace.
-
-        Args:
-            element: The element contained in the Namespace.
-            old_owner: The old owner containing the element, optional.
-
-        Returns: Gtk.TreeIter object of the model (not the sorted one!)
-        """
-
-        # Using `0` as sentinel
-        if old_owner != 0:
-            parent_iter = self.iter_for_element(old_owner)
-        elif element and element.owner:
-            parent_iter = self.iter_for_element(element.owner)
-        else:
-            parent_iter = None
-
-        if isinstance(element, UML.Relationship):
-            parent_iter = relationship_iter(self.model, parent_iter)
-
-        child_iter = self.model.iter_children(parent_iter)
-        while child_iter:
-            if self.model.get_value(child_iter, 0) is element:
-                return child_iter
-            child_iter = self.model.iter_next(child_iter)
-        return None
-
-    def _visible(self, element):
-        return isinstance(
-            element, (UML.Relationship, UML.NamedElement)
-        ) and not isinstance(
-            element, (UML.InstanceSpecification, UML.OccurrenceSpecification)
-        )
-
-    def _add(self, element, iter=None):
-        if self._visible(element):
-            if isinstance(element, UML.Relationship):
-                iter = relationship_iter(self.model, iter)
-            child_iter = self.model.append(iter, [element])
-            for e in element.ownedElement:
-                # check if owned element is indeed within parent's owner
-                # This is important since we should be able to traverse this relation both ways
-                if element is e.owner:
-                    self._add(e, child_iter)
-
-    def _remove(self, iter):
-        if iter:
-            parent_iter = self.model.iter_parent(iter)
-            self.model.remove(iter)
-            if (
-                parent_iter
-                and not self.model.iter_has_child(parent_iter)
-                and self.model.get_value(parent_iter, 0) is RELATIONSHIPS
-            ):
-                self.model.remove(parent_iter)
-
-    @event_handler(ModelReady, ModelFlushed)
-    def refresh(self, event=None):
-        """Load a new model completely."""
-        log.info("Rebuilding namespace model")
-
-        self.model.clear()
-
-        toplevel = self.element_factory.select(
-            lambda e: self._visible(e) and not e.owner
-        )
-
-        for element in toplevel:
-            if self._visible(element):
-                self._add(element)
-
-        self.event_manager.handle(NamespaceModelRefreshed(self))
-
-    @event_handler(ElementCreated)
-    def _on_element_create(self, event: ElementCreated):
-        element = event.element
-        if self._visible(element) and not self.iter_for_element(element):
-            iter = self.iter_for_element(element.owner)
-            self._add(element, iter)
-
-    @event_handler(ElementDeleted)
-    def _on_element_delete(self, event: ElementDeleted):
-        element = event.element
-        iter = self.iter_for_element(element)
-        self._remove(iter)
-
-    @event_handler(DerivedSet)
-    def _on_association_set(self, event: DerivedSet):
-        if event.property is not UML.Element.owner:
-            return
-        old_value, new_value = event.old_value, event.new_value
-
-        element = event.element
-        old_iter = self.iter_for_element(element, old_owner=old_value)
-        self._remove(old_iter)
-
-        if self._visible(element):
-            new_iter = self.iter_for_element(new_value)
-            # Should be either set (sub node) or unset (root node)
-            if bool(new_iter) == bool(new_value):
-                self._add(element, new_iter)
-
-    @event_handler(AttributeUpdated)
-    def _on_attribute_change(self, event: AttributeUpdated):
-        """Element changed, update appropriate row."""
-        if (
-            event.property is UML.Classifier.isAbstract
-            or event.property is UML.BehavioralFeature.isAbstract
-            or event.property is UML.NamedElement.name
-        ):
-            element = event.element
-
-            iter = self.iter_for_element(element)
-            if iter:
-                path = self.model.get_path(iter)
-                self.model.row_changed(path, iter)
 
 
 class Namespace(UIComponent):
